fix: move socket classes to own directory
